requirements.txt
captum==0.8.0
hydra-core==1.3.2
matplotlib==3.10.8
pandas==3.0.0
scipy==1.17.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
utils==1.0.2
